Rename: 각 모듈별 최상위 디렉토리명 변경
