--- module-api/src/main/kotlin/finn/api/ModuleApiApplication.kt
@@ -1,11 +1,0 @@
-package finn.api
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class ModuleApiApplication
-
-fun main(args: Array<String>) {
-    runApplication<ModuleApiApplication>(*args)
-}
