--- module-api/src/main/kotlin/finn/api/exception/ExceptionResponseCodeMapper.kt
@@ -1,16 +1,0 @@
-package finn.api.exception
-
-import finn.api.response.ResponseCode
-import org.springframework.stereotype.Component
-
-@Component
-object ExceptionResponseCodeMapper {
-    fun mapResponseCode(e: Exception): ResponseCode {
-        val className = e::class.simpleName ?: ResponseCode.INTERNAL_SERVER_ERROR.toString()
-
-        return ResponseCode.entries.firstOrNull { code ->
-            val keyword = code.name.replace("_", "") // 언더스코어 표기 무시
-            className.contains(keyword, true)
-        } ?: ResponseCode.INTERNAL_SERVER_ERROR
-    }
-}
