--- module-api/src/main/kotlin/finn/api/response/ErrorResponse.kt
@@ -1,4 +1,0 @@
-package finn.api.response
-
-class ErrorResponse(val code: String, message: String) {
-}
