--- module-api/src/main/kotlin/finn/api/response/SuccessResponse.kt
@@ -1,5 +1,0 @@
-package finn.api.response
-
-class SuccessResponse<T> (val code: String, message: String, val data: T?) {
-    constructor(code: String, message: String) : this(code, message, null)
-}
