--- module-api/src/main/kotlin/finn/api/response/graph/TickerGraphResponse.kt
@@ -1,13 +1,0 @@
-package finn.api.response.graph
-
-import java.math.BigDecimal
-
-data class TickerGraphResponse(
-    val period: String,
-    val graphData : List<TickerGraphResponse>
-) {
-    data class TickerGraphDataResponse(
-        val date : String,
-        val price : BigDecimal
-    )
-}
