--- module-api/src/main/kotlin/finn/api/response/search/TickerSearchPreviewListResponse.kt
@@ -1,14 +1,0 @@
-package finn.api.response.search
-
-import java.util.*
-
-data class TickerSearchPreviewListResponse(
-    val tickerSearchList: List<TickerSearchPreviewResponse>
-) {
-    data class TickerSearchPreviewResponse(
-        val tickerId: UUID,
-        val tickerCode: String,
-        val shortCompanyName: String,
-        val fullCompanyName: String
-    )
-}
