+++ module-api/src/main/kotlin/finn/moduleApi/ModuleApiApplication.kt
@@ -1,0 +1,11 @@
+package finn.moduleApi
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class ModuleApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<ModuleApiApplication>(*args)
+}
