+++ module-api/src/main/kotlin/finn/moduleApi/exception/ExceptionResponseCodeMapper.kt
@@ -1,0 +1,16 @@
+package finn.moduleApi.exception
+
+import finn.moduleApi.response.ResponseCode
+import org.springframework.stereotype.Component
+
+@Component
+object ExceptionResponseCodeMapper {
+    fun mapResponseCode(e: Exception): ResponseCode {
+        val className = e::class.simpleName ?: ResponseCode.INTERNAL_SERVER_ERROR.toString()
+
+        return ResponseCode.entries.firstOrNull { code ->
+            val keyword = code.name.replace("_", "") // 언더스코어 표기 무시
+            className.contains(keyword, true)
+        } ?: ResponseCode.INTERNAL_SERVER_ERROR
+    }
+}
