+++ module-api/src/main/kotlin/finn/moduleApi/response/ErrorResponse.kt
@@ -1,0 +1,4 @@
+package finn.moduleApi.response
+
+class ErrorResponse(val code: String, message: String) {
+}
