+++ module-api/src/main/kotlin/finn/moduleApi/response/SuccessResponse.kt
@@ -1,0 +1,5 @@
+package finn.moduleApi.response
+
+class SuccessResponse<T> (val code: String, message: String, val data: T?) {
+    constructor(code: String, message: String) : this(code, message, null)
+}
