+++ module-api/src/main/kotlin/finn/moduleApi/response/graph/TickerGraphResponse.kt
@@ -1,0 +1,13 @@
+package finn.moduleApi.response.graph
+
+import java.math.BigDecimal
+
+data class TickerGraphResponse(
+    val period: String,
+    val graphData : List<TickerGraphResponse>
+) {
+    data class TickerGraphDataResponse(
+        val date : String,
+        val price : BigDecimal
+    )
+}
