+++ module-api/src/main/kotlin/finn/moduleApi/response/news/NewsListResponse.kt
@@ -1,0 +1,23 @@
+package finn.moduleApi.response.news
+
+import java.util.*
+
+data class NewsListResponse(
+    val total: Int,
+    val newsList: List<NewsDataResponse>,
+    val pageNumber: Int,
+    val hasNext: Boolean
+) {
+    data class NewsDataResponse(
+        val newsId: UUID,
+        val title: String,
+        val description: String,
+        val shortCompanyName: String,
+        val thumbnailUrl: String,
+        val contentUrl: String,
+        val publishedDate: String,
+        val source: String,
+        val sentiment: String,
+        val reasoning: String? = null
+    )
+}
