+++ module-api/src/main/kotlin/finn/moduleApi/response/search/TickerSearchPreviewListResponse.kt
@@ -1,0 +1,14 @@
+package finn.moduleApi.response.search
+
+import java.util.*
+
+data class TickerSearchPreviewListResponse(
+    val tickerSearchList: List<TickerSearchPreviewResponse>
+) {
+    data class TickerSearchPreviewResponse(
+        val tickerId: UUID,
+        val tickerCode: String,
+        val shortCompanyName: String,
+        val fullCompanyName: String
+    )
+}
