--- module-api/src/test/kotlin/finn/api/ModuleApiApplicationTests.kt
@@ -1,13 +1,0 @@
-package finn.api
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ModuleApiApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
