+++ module-api/src/test/kotlin/finn/moduleApi/ModuleApiApplicationTests.kt
@@ -1,0 +1,13 @@
+package finn.moduleApi
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class ModuleApiApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
