--- module-common/src/main/kotlin/finn/common/exception/CommonException.kt
@@ -1,11 +1,0 @@
-package finn.common.exception
-
-abstract class CommonException(
-    message: String,
-    cause: Throwable?
-) : RuntimeException(message, cause) {
-
-    constructor(message: String) : this(
-        message, null
-    )
-}
