+++ module-common/src/main/kotlin/finn/moduleCommon/exception/CommonException.kt
@@ -1,0 +1,11 @@
+package finn.moduleCommon.exception
+
+abstract class CommonException(
+    message: String,
+    cause: Throwable?
+) : RuntimeException(message, cause) {
+
+    constructor(message: String) : this(
+        message, null
+    )
+}
